requirements-only-tests-locked.txt
# This file was autogenerated by uv via the following command:
#    uv export -o requirements-only-tests-locked.txt --no-hashes --no-dev --no-emit-project --only-group tests
asttokens==3.0.0
colorama==0.4.6 ; sys_platform == 'win32'
coverage==7.6.10
decorator==5.1.1
exceptiongroup==1.2.2 ; python_full_version < '3.11'
executing==2.1.0
flexcache==0.3
flexparser==0.4
globalwarmingpotentials==0.11.1
iniconfig==2.0.0
ipython==8.18.1 ; python_full_version < '3.10'
ipython==8.31.0 ; python_full_version >= '3.10'
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==2.0.2 ; python_full_version < '3.10'
numpy==2.2.2 ; python_full_version >= '3.10'
openscm-units==0.6.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0 ; (python_full_version < '3.10' and sys_platform == 'emscripten') or (sys_platform != 'emscripten' and sys_platform != 'win32')
pint==0.24.4
platformdirs==4.3.6
pluggy==1.5.0
prompt-toolkit==3.0.49
ptyprocess==0.7.0 ; (python_full_version < '3.10' and sys_platform == 'emscripten') or (sys_platform != 'emscripten' and sys_platform != 'win32')
pure-eval==0.2.3
pygments==2.19.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-datadir==1.5.0
pytest-regressions==2.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyyaml==6.0.2
scipy==1.13.1 ; python_full_version < '3.10'
scipy==1.15.1 ; python_full_version >= '3.10'
six==1.17.0
stack-data==0.6.3
tomli==2.2.1 ; python_full_version <= '3.11'
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
